% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{build_ppi_network}
\alias{build_ppi_network}
\title{Retrive a desired PPI network from STRINGdb in a PPIpriority object}
\usage{
build_ppi_network(version ="11.5", species = 9606,
score_threshold=400, input_directory=getwd())
}
\arguments{
\item{version}{a character value, identifies the version of STRINGdb to use.}

\item{species}{a numeric value, that identifies the species of interest.}

\item{score_threshold}{a numeric value set to prune edges in the PPI network,
based on the level of confidence in the interaction. It should be a numeric
value between 0 and 1000. Values near to 1000 may cause problems with
prioritization, since will cause an intense pruning of the network.

Note: score does not indicate the strength or the specificity of the
interaction. Instead, it is an indicator of confidence.}

\item{input_directory}{Directory in which STRINGdb will download and
store the network information.}
}
\value{
The output is an a PPIpriority_obj object (S4 class). The
object stores the STRINGdb connection and the PPI graph, that can
be accessed with the methods "get_STRINGdb_connection" and
"get_network", respectively.
The network is unweighted and the edges are already pruned for the
input value of score threshold. The vertices in this network have
also the attribute "preferred name" which is the HUGO nomeclature
for the STRING id in human network.

Please refer to STRINGdb for version and species accessors.
If the function fails to download the required files because of having
reached download timeout, please extend timeout using "options(timeout=600)".
Removed the uncompleted downloaded files and call this function again.
}
\description{
The build_ppi_network function is used to retrieve and
build a PPI network from the STRINGdb in a PPIpriority object. This
is the first step of the PPIpriority workflow, which allows to
perform random walk based target prioritization by calling
"prioritize_targets" on the output of this function.
}
\examples{
#Set timeout, not necessary when using the package
options(timeout=600)

# Retrieve the PPI netowrk using STRINGdb
network_ppi_obj<- build_ppi_network(version ="11.5", species = 9606,
score_threshold=800, input_directory=getwd()) #Build a human PPI
}
