% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation_functions.R
\name{check_gene_names_input}
\alias{check_gene_names_input}
\title{Check gene names input}
\usage{
check_gene_names_input(input_gene_names)
}
\arguments{
\item{input_gene_names}{A vector of gene names}
}
\value{
An appropriate error message
}
\description{
Input validation function: called during the input validation
step of "prioritize_targets".
}
