% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation_functions.R
\name{check_input_threshold}
\alias{check_input_threshold}
\title{Warning for high score_threshold}
\usage{
check_input_threshold(threshold)
}
\arguments{
\item{threshold}{A vector of gene names}
}
\value{
An appropriate error message
}
\description{
Input validation function: called during the input validation
step of "prioritize_targets".
}
