% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation_functions.R
\name{check_mapping_output_candidates}
\alias{check_mapping_output_candidates}
\title{Check mapping output for candidates}
\usage{
check_mapping_output_candidates(df, input_gene_names)
}
\arguments{
\item{df}{A data.frame containing the columns: "gene_names" , "STRING_id"}

\item{input_gene_names}{A vector of gene names}
}
\value{
An appropriate error message or a complementary message to help
the user understanding which are the unmapped gene names.
}
\description{
Input validation function: called after mapping of input gene
names (candidate) in "prioritize_targets". Prevents
"prioritize_targets" to continue if input genes can not be mapped to the
STRING ids of the PPI network. It also informs the user which are the genes
that are excluded from the analysis because of unsuccessful mapping.
The warning is raised by STRINGdb map.
}
