% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation_functions.R
\name{check_numeric_input}
\alias{check_numeric_input}
\title{Check numeric inputs}
\usage{
check_numeric_input(x)
}
\arguments{
\item{x}{an input parameter to check}
}
\value{
An appropriate error message
}
\description{
Input validation function: called during the input validation
step of "prioritize_targets"to check RWR inputs.
}
