% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation_functions.R
\name{check_PPIpriority_obj_input}
\alias{check_PPIpriority_obj_input}
\title{Check PPI priority object input}
\usage{
check_PPIpriority_obj_input(obj_input)
}
\arguments{
\item{obj_input}{A PPIpriority_obj}
}
\value{
An appropriate error message
}
\description{
Input validation function: called during the input validation
step of "prioritize_targets".
}
