% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation_functions.R
\name{check_presence_cand_in_res}
\alias{check_presence_cand_in_res}
\title{Check presence for candidates in the RW results}
\usage{
check_presence_cand_in_res(can_df, results)
}
\arguments{
\item{can_df}{A data.frame of mapped candidate ids from STRINGdb
containing the columns: "gene_names" , "STRING_id"}

\item{results}{A data.frame, internal result of prioritize_targets.}
}
\value{
An appropriate warning
}
\description{
Internal validation function:  warns that candidates are not
in RW results.
}
