% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation_functions.R
\name{check_presence_seeds_in_graph}
\alias{check_presence_seeds_in_graph}
\title{Check presence for seeds in the graph}
\usage{
check_presence_seeds_in_graph(seed_df, adj_matrix)
}
\arguments{
\item{seed_df}{A data.frame of mapped seed ids from STRINGdb
containing the columns: "gene_names" , "STRING_id"}

\item{adj_matrix}{A matrix, adj. matrix from igraph obj in ppi_priority obj}
}
\value{
An appropriate error message or a complementary message to help
the user understanding which are the missing seeds.
}
\description{
Internal validation function:  Prevents
"prioritize_targets" to continue if mapped input gene id can not be found
in the PPI network.
}
