% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{get_STRINGdb_connection}
\alias{get_STRINGdb_connection}
\title{Get PPI network from a PPIpriority_obj}
\usage{
get_STRINGdb_connection(object)
}
\arguments{
\item{object}{a PPIpriority object.}
}
\value{
A STRINGdb object.
}
\description{
Accessor to the STRINGdb obj stored in a PPIpriority obj.
}
\examples{
#Set extended timeout, not necessary while using the package
options(timeout=600)

#Build human PPI
network_ppi_obj<- build_ppi_network(version ="11.5", species = 9606,
score_threshold=800, input_directory=getwd())

# Access STRINGdb
get_STRINGdb_connection(network_ppi_obj)
}
