% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\docType{class}
\name{PPIpriority_obj-class}
\alias{PPIpriority_obj-class}
\title{PPI priority object}
\description{
PPI priority object. It stores STRINGdb and igraph objects.
}
\section{Slots}{

\describe{
\item{\code{STRINGdb_connection}}{A STRINGdb object}

\item{\code{graph}}{An igraph object}
}}

\examples{
#Set timeout, not necessary while using the package
options(timeout=600)

#Build a human PPI
network_ppi_obj<- build_ppi_network(version ="11.5", species = 9606,
score_threshold=800, input_directory=getwd())

# Access network igraph
get_network(network_ppi_obj)

# Access STRINGdb
get_STRINGdb_connection(network_ppi_obj)
}
