% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\docType{methods}
\name{get_network,PPIpriority_obj-method}
\alias{get_network,PPIpriority_obj-method}
\alias{get_STRINGdb_connection,PPIpriority_obj-method}
\title{Get PPI network from a PPIpriority_obj}
\usage{
\S4method{get_network}{PPIpriority_obj}(object)

\S4method{get_STRINGdb_connection}{PPIpriority_obj}(object)
}
\arguments{
\item{object}{a PPIpriority object.}
}
\value{
An igraph object

A STRINGdb object
}
\description{
Accessor to the igraph object stored in a PPIpriority obj.

Accessor to the STRINGdb obj stored in a PPIpriority obj.
}
\examples{
#Set extended timeout, not necessary while using the package
options(timeout=600)

#Build human PPI
network_ppi_obj<- build_ppi_network(version ="11.5", species = 9606,
score_threshold=800, input_directory=getwd())

# Access network igraph
get_network(network_ppi_obj)
#Set extended timeout, not necessary while using the package
options(timeout=600)

#Build a human PPI
network_ppi_obj<- build_ppi_network(version ="11.5", species =9606,
score_threshold=800, input_directory=getwd())

# Access STRINGdb
get_STRINGdb_connection(network_ppi_obj)
}
