% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{prioritize_targets}
\alias{prioritize_targets}
\title{Prioritize targets on PPI network via random walk from seed position(s).}
\usage{
prioritize_targets(PPIpriority_obj, seeds, candidates=NA, gamma= 0.6,
 tmax= 1000, eps=1e-10)
}
\arguments{
\item{PPIpriority_obj}{A PPIpriority object.}

\item{seeds}{A vector of characters describing genes of interest
for seeding the random walk.}

\item{candidates}{A vector of characters describing genes which may
be associated to the phenotype of interest.}

\item{gamma}{A numeric value, restart parameter.}

\item{tmax}{A numeric value, maximum number of iterations.}

\item{eps}{A numeric value, maximum allowed difference between the
computed probabilities at the steady state.}
}
\value{
data.frame of the target prioritization.
If the function fails to download the required files for downloading
matching annotations because of having reached download timeout,
please extend it using options(timeout=600)
}
\description{
Target prioritization based on Random Walk (RW) in the
selected STRINGdb PPI network based on seed potion(s).The function
requires to set as input seed genes and a STRINGdb PPI network,
which can be conveniently built using the function "build_ppi_network"
available in this package. Random Walk is based on the RWR function from
the RANKS package.
}
\examples{
#Set extended timeout, not necessary while using the package
options(timeout=600)

# Build a human PPI
network_ppi_obj<- build_ppi_network(version ="11.5", species = 9606,
score_threshold=800, input_directory=getwd())

# Run the random walk using ERP44 and PRDX4 as seeds, print the output
prioritize_targets(network_ppi_obj, seeds=c("ERP44", "PRDX4"))
}
\references{
Wang et al. (2011) Network-based methods for human disease
gene prediction. Briefings in Functional Genomics 10(5):280–293.
}
